+++ packages/sdk/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## NEXT
+
+### Changed
+
+- migrated subgraphs default base URL (<https://thegraph.iex.ec>)
+
 ## [2.0.0-beta.15] (2025-04-04)
 
 ### Changed